make countdown a reusable component
