+++ README.md
@@ -14,7 +14,7 @@ Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
 The page will reload when you make changes.\
 You may also see any lint errors in the console.
 
-## Countdown & World Clock Application
+<!-- ## Countdown & World Clock Application
 
 The Countdown & World Clock Application is a React-based application that provides two main features:
 
@@ -48,4 +48,4 @@ World Clock Component
 The World Clock component is responsible for displaying the current time in various time zones around the world. The specific implementation details of World Clock are not shown in the provided code, but it would typically involve using JavaScript's Date object and possibly some external library to handle time zones.
 
 Usage
-The application can be used in any React application by importing the Home component and using it in the render method of another component or in the return statement of a functional component.
+The application can be used in any React application by importing the Home component and using it in the render method of another component or in the return statement of a functional component. -->
